Add site-wide library
config.ts is a collection of text and paths used throughout the site.
The purpose of this library is to reduce the number of files that need to be changed, so that when changes are made to the site, changes to config.ts will result in changes to how the site is displayed.